specialisation_nan/requirements.txt
Django==2.2.7
Pillow==6.2.1
django-admin-interface==0.11.2
django-tinymce4-lite==1.7.5
django-filebrowser-no-grappelli==3.7.8
django-filter==2.2.0
graphene==2.1.8
graphene-django==2.6.0
django-admin-generator
psycopg2-binary
django-mathfilters
requests
django-user-agents
python-memcached
APScheduler